+++ lib/get5_api/teams.ex
@@ -8,6 +8,7 @@ defmodule Get5Api.Teams do
 
   alias Get5Api.Teams.Team
   alias Get5Api.Teams.Player
+  alias Get5Api.Accounts.User
 
   @doc """
   Returns the list of teams.
@@ -18,9 +19,24 @@ defmodule Get5Api.Teams do
       [%Team{}, ...]
 
   """
-  def list_teams do
+  def list_teams(user_id) do
+    if user_id do
+      Repo.all(
+        from t in Team,
+          where: t.public == true or t.user_id == ^user_id,
+          preload: [:user],
+          order_by: [asc: :inserted_at]
+      )
+    else
+      list_public_teams()
+    end
+  end
+
+  def list_public_teams() do
     Repo.all(
       from t in Team,
+        where: t.public == true,
+        preload: [:user],
         order_by: [asc: :inserted_at]
     )
   end
@@ -39,9 +55,9 @@ defmodule Get5Api.Teams do
       ** (Ecto.NoResultsError)
 
   """
-  def get_team!(id), do: Repo.get!(Team, id)
+  def get_team!(id), do: Repo.get!(Team, id) |> Repo.preload([:user])
 
-  def get_team(id), do: Repo.get(Team, id)
+  def get_team(id), do: Repo.get(Team, id) |> Repo.preload([:user])
 
   @doc """
   Creates a team.
@@ -73,7 +89,15 @@ defmodule Get5Api.Teams do
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_team(%Team{} = team, attrs) do
+  def update_team(%User{} = user, %Team{} = team, attrs) do
+    if team.user_id == user.id do
+      update_team(team, attrs)
+    else
+      {:error, :unauthorized}
+    end
+  end
+
+  def update_team(%Team{} = team, attrs \\ %{}) do
     team
     |> Team.changeset(attrs)
     |> Repo.update()
@@ -115,6 +139,14 @@ defmodule Get5Api.Teams do
     end
   end
 
+  def remove_player(%User{} = user, %Team{} = team, %Player{} = player) do
+    if team.user_id == user.id do
+      remove_player(team, player)
+    else
+      {:error, :unauthorized}
+    end
+  end
+
   def remove_player(%Team{} = team, %Player{} = player) do
     players = Enum.reject(team.players, &(&1.steam_id == player.steam_id))
 
@@ -145,10 +177,14 @@ defmodule Get5Api.Teams do
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_team(%Team{} = team) do
-    team
-    |> Team.changeset(%{})
-    |> Repo.delete()
+  def delete_team(%User{} = user, %Team{} = team) do
+    if team.user_id == user.id do
+      team
+      |> Team.changeset(%{})
+      |> Repo.delete()
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
